+++ .env.test.example
@@ -1,4 +1,4 @@
-# This environment variable is only used for testing
+# The environment variables are only used for testing
 API_USERNAME =
 API_PASSWORD =
 TOJ_USERNAME =
