+++ CODE_OF_CONDUCT.md
@@ -1,0 +1,29 @@
+# Code of Conduct
+
+## Introduction
+
+We are committed to providing a friendly, safe, and inclusive environment for all participants. This code of conduct outlines our expectations for everyone involved in our community, as well as the consequences for unacceptable behavior.
+
+## Expected Behavior
+
+- Be respectful and considerate towards others.
+- Use inclusive language and avoid offensive or discriminatory comments.
+- Be open to constructive feedback and different perspectives.
+- Exercise empathy and understanding towards others.
+
+## Unacceptable Behavior
+
+The following behaviors are considered unacceptable and will not be tolerated:
+
+- Harassment, discrimination, or any form of offensive behavior.
+- Intimidation, threats, or personal attacks.
+- Disruptive or disrespectful comments or actions.
+- Any other behavior that creates an unsafe or unwelcoming environment.
+
+## Reporting and Enforcement
+
+If you witness or experience any behavior that violates this code of conduct, please report it to [2210227279@qq.com]. All reports will be reviewed and investigated promptly. We reserve the right to take appropriate action, including but not limited to warning, temporary or permanent ban, or legal action.
+
+## Conclusion
+
+By participating in our community, you agree to abide by this code of conduct. We expect all participants to uphold these standards and create a positive and inclusive environment for everyone.
