+++ src/urls/card.ts
@@ -1,0 +1,13 @@
+export const CARD_I_VERSION_URL = 'https://webvpn.tsinghua.edu.cn/http/77726476706e69737468656265737421f1e751d2242326446d0187ab9040227b239c069338e2/api/CardIVersion'
+
+export const CARD_USER_BY_TOKEN_URL = 'https://card.tsinghua.edu.cn/login/getUserInfoFromToken'
+export const CARD_INFO_BY_USER_URL = 'https://card.tsinghua.edu.cn/business/getCardUserinfo'
+export const CARD_PHOTO_URL = 'https://card.tsinghua.edu.cn/myaccount/showDbImage?idserial='
+export const CARD_TRANSACTION_URL = 'https://card.tsinghua.edu.cn/business/querySelfTradeList'
+export const CARD_CHANGE_PWD_URL = 'https://card.tsinghua.edu.cn/business/modifyPwdByPhoneVerify'
+export const CARD_MOD_MAX_CONSUME_URL = 'https://card.tsinghua.edu.cn/business/modifyCardMaxConsamt'
+export const CARD_REPORT_LOSS_URL = 'https://card.tsinghua.edu.cn/business/cardReportLoss'
+export const CARD_CANCEL_LOSS_URL = 'https://card.tsinghua.edu.cn/business/solutionHang'
+export const CARD_RECHARGE_FROM_BANK_URL = 'https://card.tsinghua.edu.cn/business/moblieRecharge'
+export const CARD_RECHARGE_FROM_WECHAT_URL = 'https://card.tsinghua.edu.cn/wx/rechard/getRechargeHtml'
+export const CARD_RECHARGE_FROM_ALIPAY_URL = 'https://card.tsinghua.edu.cn/wx/rechard/qrcode'
