+++ src/urls/gitlab.ts
@@ -1,0 +1,3 @@
+export const GITLAB_LOGIN_URL = 'https://git.tsinghua.edu.cn/users/sign_in'
+export const GITLAB_AUTH_URL = 'https://git.tsinghua.edu.cn/users/auth/thuid'
+export const GITLAB_API_BASE_URL = 'https://git.tsinghua.edu.cn/api/v4'
